klantenservice_applicatie/requirements-bigquery.txt
google-cloud-bigquery>=3.11.4
google-auth>=2.23.0
python-dotenv>=1.0.0 